+++ rust/gitxetcore/src/diff/twb.rs
@@ -1,28 +1,18 @@
-use tableau_summary::twb::diff::{TwbDiffProcessor};
+use tableau_summary::twb::diff::schema::TwbSummaryDiffContent;
+use tableau_summary::twb::diff::util::DiffProducer;
 use tableau_summary::twb::TwbSummary;
-use crate::diff::output::SummaryDiffData;
+
 use crate::diff::{DiffError, SummaryDiffProcessor};
+use crate::diff::output::SummaryDiffData;
 use crate::diff::output::SummaryDiffData::Twb;
 use crate::summaries::{FileSummary, SummaryType};
-
 
 /// Processes diffs of Twb Summaries, taking in [TwbSummary]s and producing a [TwbSummaryDiffContent]
 /// indicating the delta between them.
 ///
 /// Currently, this will just provide the two summaries (before + after). In the future, we can
 /// change this to become more intelligent (actually identify what changed between the workbooks).
-pub struct TwbSummaryDiffProcessor {
-    proc: TwbDiffProcessor
-}
-
-impl TwbSummaryDiffProcessor {
-    pub fn new() -> Self {
-        TwbSummaryDiffProcessor {
-            proc: TwbDiffProcessor {},
-        }
-    }
-}
-
+pub struct TwbSummaryDiffProcessor {}
 
 impl SummaryDiffProcessor for TwbSummaryDiffProcessor {
     type SummaryData = TwbSummary;
@@ -37,15 +27,15 @@ impl SummaryDiffProcessor for TwbSummaryDiffProcessor {
 
     fn get_data<'a>(&'a self, summary: &'a FileSummary) -> Option<&Self::SummaryData> {
         summary.additional_summaries.as_ref()
-            .and_then(|ext|ext.twb.as_ref())
+            .and_then(|ext| ext.twb.as_ref())
     }
 
     fn get_insert_diff(&self, summary: &TwbSummary) -> Result<SummaryDiffData, DiffError> {
-        Ok(Twb(self.proc.get_insert_diff(summary)?))
+        Ok(Twb(TwbSummaryDiffContent::new_addition(summary)))
     }
 
     fn get_remove_diff(&self, summary: &TwbSummary) -> Result<SummaryDiffData, DiffError> {
-        Ok(Twb(self.proc.get_remove_diff(summary)?))
+        Ok(Twb(TwbSummaryDiffContent::new_deletion(summary)))
     }
 
     fn get_diff_impl(
@@ -53,6 +43,6 @@ impl SummaryDiffProcessor for TwbSummaryDiffProcessor {
         before: &TwbSummary,
         after: &TwbSummary,
     ) -> Result<SummaryDiffData, DiffError> {
-        Ok(Twb(self.proc.get_diff_impl(before, after)?))
+        Ok(Twb(TwbSummaryDiffContent::new_diff(before, after)))
     }
 }
